requirements.txt
PyPDF2
PyMuPDF
pdfplumber
pandas
tabulate
openai
pillow
requests